# [Misc] Add tuned R1 w8a8 and MoE configs for NVIDIA L20