Reopen https://github.com/vllm-project/vllm/pull/15117 and https://github.com/vllm-project/vllm/pull/15126 as one completed PR here. 

Hi~, we are the tech team from vipshop.com and mainly use vLLM in our production deployment. This PR is mainly focused on optimizing DeepSeek-R1 671B inference performance on NVIDIA L20 devices using vLLM with pp3 + tp8. We hope that this PR will further enhance the performance of vLLM on NVIDIA L20. This PR has the following contributions:

- 1. Tuned fused MoE and w8a8 block fp8 performance on NVIDIA L20

Added tuned w8a8 and fused MoE configurations for NVIDIA L20, which can boost the decode throughput of DeepSeek-R1 671B with pp3 + tp8 by up to `10~15%`.

- 2. Add w8a8 block fp8 tuning script 

Adapted from https://github.com/sgl-project/sglang/blob/main/benchmark/kernels/quantization/tuning_block_wise_kernel.py.

```bash
# Tune triton w8a8 block fp8 for DeepSeek-V3/DeepSeek-R1:
python3 benchmark_w8a8_block_fp8.py --tp-size 8 --input-type fp8
# Then copy then configs to vllm/model_executor/layers/quantization/utils/configs
```

Before this PR:  

```bash
WARNING 03-19 15:21:07 [fp8_utils.py:431] Using default W8A8 Block FP8 kernel config. Performance might be sub-optimal! Config file not found at /usr/local/lib/python3.10/dist-packages/vllm/model_executor/layers/quantization/utils/configs/N=1536,K=7168,device_name=NVIDIA_L20,dtype=fp8_w8a8,block_shape=[128,128].json
........
WARNING 03-19 15:21:08 [fused_moe.py:874] Using default MoE config. Performance might be sub-optimal! Config file not found at /usr/local/lib/python3.10/site-packages/vllm/model_executor/layers/fused_moe/configs/E=256,N=256,device_name=NVIDIA_L20,dtype=fp8_w8a8,block_shape=[128,128].json

```

After this PR:

```bash
INFO 03-22 12:10:07 [fp8_utils.py:422] Using configuration from /usr/local/lib/python3.10/dist-packages/vllm/model_executor/layers/quantization/utils/configs/N=1536,K=7168,device_name=NVIDIA_L20,dtype=fp8_w8a8,block_shape=[128,128].json for W8A8 Block FP8 kernel.
INFO 03-22 12:10:07 [fp8_utils.py:422] Using configuration from /usr/local/lib/python3.10/dist-packages/vllm/model_executor/layers/quantization/utils/configs/N=576,K=7168,device_name=NVIDIA_L20,dtype=fp8_w8a8,block_shape=[128,128].json for W8A8 Block FP8 kernel.
INFO 03-22 12:10:07 [fp8_utils.py:422] Using configuration from /usr/local/lib/python3.10/dist-packages/vllm/model_executor/layers/quantization/utils/configs/N=3072,K=1536,device_name=NVIDIA_L20,dtype=fp8_w8a8,block_shape=[128,128].json for W8A8 Block FP8 kernel.
INFO 03-22 12:10:07 [fp8_utils.py:422] Using configuration from /usr/local/lib/python3.10/dist-packages/vllm/model_executor/layers/quantization/utils/configs/N=7168,K=2048,device_name=NVIDIA_L20,dtype=fp8_w8a8,block_shape=[128,128].json for W8A8 Block FP8 kernel.
INFO 03-22 12:10:07 [fp8_utils.py:422] Using configuration from /usr/local/lib/python3.10/dist-packages/vllm/model_executor/layers/quantization/utils/configs/N=4608,K=7168,device_name=NVIDIA_L20,dtype=fp8_w8a8,block_shape=[128,128].json for W8A8 Block FP8 kernel.
INFO 03-22 12:10:07 [fp8_utils.py:422] Using configuration from /usr/local/lib/python3.10/dist-packages/vllm/model_executor/layers/quantization/utils/configs/N=7168,K=2304,device_name=NVIDIA_L20,dtype=fp8_w8a8,block_shape=[128,128].json for W8A8 Block FP8 kernel.
INFO 03-22 12:10:07 [fp8_utils.py:422] Using configuration from /usr/local/lib/python3.10/dist-packages/vllm/model_executor/layers/quantization/utils/configs/N=512,K=7168,device_name=NVIDIA_L20,dtype=fp8_w8a8,block_shape=[128,128].json for W8A8 Block FP8 kernel.
INFO 03-22 12:10:07 [fp8_utils.py:422] Using configuration from /usr/local/lib/python3.10/dist-packages/vllm/model_executor/layers/quantization/utils/configs/N=7168,K=256,device_name=NVIDIA_L20,dtype=fp8_w8a8,block_shape=[128,128].json for W8A8 Block FP8 kernel.
INFO 03-22 12:10:08 [fused_moe.py:874] Using configuration from /usr/local/lib/python3.10/dist-packages/vllm/model_executor/layers/fused_moe/configs/E=256,N=256,device_name=NVIDIA_L20,dtype=fp8_w8a8,block_shape=[128,128].json for MoE layer.
```

Performance 4K input & 1K output:

```bash
#### command
python3 benchmark_serving.py \
        --backend vllm \
        --model /workspace/DeepSeek-R1 \
        --port 8862 \
        --endpoint /v1/completions \
        --dataset-name random \
        --dataset-path ${SHAREGPT_DATASET_PATH} \
        --random-input-len 4096 \
        --random-output-len 1024 \
        --random-prefix-len 0 \
        --ignore-eos \
        --max-concurrency 16 \
        --num-prompts 16

#### Before
# 131.55 TPS with concurrency 16, 4K input & 1K output, TTFT 8749ms
Maximum request concurrency: 16
============ Serving Benchmark Result ============
Successful requests:                     16
Benchmark duration (s):                  124.74
Total input tokens:                      65536
Total generated tokens:                  16384
Request throughput (req/s):              0.13
Output token throughput (tok/s):         131.35
Total Token throughput (tok/s):          656.75
---------------Time to First Token----------------
Mean TTFT (ms):                          8749.86
Median TTFT (ms):                        9009.79
P99 TTFT (ms):                           13242.30
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          112.99
Median TPOT (ms):                        112.77
P99 TPOT (ms):                           117.74
---------------Inter-token Latency----------------
Mean ITL (ms):                           112.99
Median ITL (ms):                         108.20
P99 ITL (ms):                            123.80
==================================================

#### After 
# 145.91 TPS with concurrency 16, 4K input & 1K output, TTFT 8155ms
Maximum request concurrency: 16
============ Serving Benchmark Result ============
Successful requests:                     16
Benchmark duration (s):                  112.29
Total input tokens:                      65536
Total generated tokens:                  16384
Request throughput (req/s):              0.14
Output token throughput (tok/s):         145.91
Total Token throughput (tok/s):          729.54
---------------Time to First Token----------------
Mean TTFT (ms):                          8155.07
Median TTFT (ms):                        8429.93
P99 TTFT (ms):                           12540.61
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          101.42
Median TPOT (ms):                        101.19
P99 TPOT (ms):                           106.09
---------------Inter-token Latency----------------
Mean ITL (ms):                           101.42
Median ITL (ms):                         96.11
P99 ITL (ms):                            113.22
==================================================
```
 
Performance 1K input & 4K output

```bash
#### command
python3 benchmark_serving.py \
        --backend vllm \
        --model /workspace/DeepSeek-R1 \
        --port 8862 \
        --endpoint /v1/completions \
        --dataset-name random \
        --dataset-path ${SHAREGPT_DATASET_PATH} \
        --random-input-len 1024 \
        --random-output-len 4096 \
        --random-prefix-len 0 \
        --ignore-eos \
        --max-concurrency 32 \
        --num-prompts 32

#### Before
# 250.71 TPS with concurrency 32, 1K input & 4K output, TTFT 4367ms
Maximum request concurrency: 32
============ Serving Benchmark Result ============
Successful requests:                     32
Benchmark duration (s):                  522.81
Total input tokens:                      32768
Total generated tokens:                  131072
Request throughput (req/s):              0.06
Output token throughput (tok/s):         250.71
Total Token throughput (tok/s):          313.39
---------------Time to First Token----------------
Mean TTFT (ms):                          4367.81
Median TTFT (ms):                        4440.48
P99 TTFT (ms):                           6910.16
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          126.53
Median TPOT (ms):                        126.51
P99 TPOT (ms):                           127.27
---------------Inter-token Latency----------------
Mean ITL (ms):                           126.53
Median ITL (ms):                         125.66
P99 ITL (ms):                            145.68
==================================================

#### After
# 295.71 TPS with concurrency 32, 1K input & 4K output, TTFT 4152.47ms
Maximum request concurrency: 32
============ Serving Benchmark Result ============
Successful requests:                     32
Benchmark duration (s):                  443.24
Total input tokens:                      32768
Total generated tokens:                  131072
Request throughput (req/s):              0.07
Output token throughput (tok/s):         295.71
Total Token throughput (tok/s):          369.64
---------------Time to First Token----------------
Mean TTFT (ms):                          4152.47
Median TTFT (ms):                        4207.50
P99 TTFT (ms):                           6710.61
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          107.16
Median TPOT (ms):                        107.15
P99 TPOT (ms):                           107.88
---------------Inter-token Latency----------------
Mean ITL (ms):                           107.16
Median ITL (ms):                         106.29
P99 ITL (ms):                            125.92
==================================================
```