@robertgshaw2-redhat 2025-03-22
I dont have an L20 to verify, but i trust your results!