@tensorflowt 2025-05-08
Thank you for your efforts—the improvements are evident! I am currently facing some challenges during implementation and would appreciate your guidance.

To provide context, my setup involves an NVIDIA RTX 4090 GPU running the Qwen3-30B-A3B-FP8 model. I attempted to generate the necessary JSON configuration files by executing:

 ``` 
python3 benchmark_w8a8_block_fp8.py --tp-size 8 --input-type fp8  
 ``` 

However, upon deploying the service, I observed a discrepancy between the expected JSON filenames and those generated. Could you help identify the potential cause of this mismatch?

json file generated by the code：

![489296e5f53ce128c52bc447c1e9f91a](https://github.com/user-attachments/assets/8ddb78d2-ad04-4aa3-bb30-8200480e1a37)

Required json file：

![5644c73883ca4ad81326edb72a5c0282](https://github.com/user-attachments/assets/92548d83-b920-414e-bf71-db08125b83af)

The inference test command of vllm is as follows：
 ``` 
CUDA_VISIBLE_DEVICES=0,1,2,3 \
VLLM_USE_V1=0 \
VLLM_USE_MODELSCOPE=true \
vllm serve /work/models/Qwen/Qwen3-30B-A3B-FP8 \
--enable-reasoning \
--reasoning-parser deepseek_r1 \
--served-model-name Qwen3-30B-A3B-FP8 \
--gpu-memory-utilization 0.7 \
--tensor-parallel-size 4  \
--trust-remote-code \
--enable-chunked-prefill \
--port 8000 \
--kv-cache-dtype fp8 \
--enable-expert-parallel
 ```